+++ src/main/java/com/tkl/refactoring/chapter1/model/EnrichedPerformance.java
@@ -1,4 +1,4 @@
-package com.tkl.refactoring.chapter1;
+package com.tkl.refactoring.chapter1.model;
 
 public record EnrichedPerformance(Performance performance, Play play, int amount, int volumeCredits) {
 }
