+++ src/main/java/com/tkl/refactoring/chapter1/model/Performance.java
@@ -1,4 +1,4 @@
-package com.tkl.refactoring.chapter1;
+package com.tkl.refactoring.chapter1.model;
 
 public record Performance(String playID, int audience) {
