+++ src/main/java/com/tkl/refactoring/chapter1/model/Play.java
@@ -1,4 +1,4 @@
-package com.tkl.refactoring.chapter1;
+package com.tkl.refactoring.chapter1.model;
 
 public record Play(String name, String type) {
